+++ test/matrix.cc
@@ -302,6 +302,48 @@ TEST(Matrix, DISABLED_test_matrix_osrm) {
   }
 }
 
+const auto test_request_partial = R"({
+    "sources":[
+      {"lat":52.103948,"lon":5.06813}
+    ],
+    "targets":[
+      {"lat":52.106126,"lon":5.101497},
+      {"lat":52.100469,"lon":5.087099},
+      {"lat":52.103105,"lon":5.081005},
+      {"lat":52.094273,"lon":5.075254}
+    ],
+    "costing":"auto",
+    "matrix_locations":2
+  })";
+
+TEST(Matrix, partial_matrix) {
+  loki_worker_t loki_worker(config);
+
+  Api request;
+  ParseApi(test_request_partial, Options::sources_to_targets, request);
+  loki_worker.matrix(request);
+  adjust_scores(*request.mutable_options());
+
+  GraphReader reader(config.get_child("mjolnir"));
+
+  sif::mode_costing_t mode_costing;
+  mode_costing[0] =
+      CreateSimpleCost(request.options().costings().find(request.options().costing_type())->second);
+
+  TimeDistanceMatrix timedist_matrix;
+  std::vector<TimeDistance> results =
+      timedist_matrix.SourceToTarget(request.options().sources(), request.options().targets(), reader,
+                                     mode_costing, sif::TravelMode::kDrive, 400000.0,
+                                     request.options().matrix_locations());
+  uint32_t found = 0;
+  for (uint32_t i = 0; i < results.size(); ++i) {
+    if (results[i].dist > 0) {
+      ++found;
+    }
+  }
+  EXPECT_EQ(found, 2) << " partial result did not find 2 results as expected";
+}
+
 int main(int argc, char* argv[]) {
   logging::Configure({{"type", ""}}); // silence logs
   testing::InitGoogleTest(&argc, argv);